backend/requirements.txt
# Backend Requirements

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
PyPDF2==3.0.1
pdf2image==1.16.3
python-multipart==0.0.6
pydantic==2.5.0
